extract Login component
