+++ src/client/src/components/Login.css
@@ -1,0 +1,30 @@
+.auth {
+  display: flex;
+  flex-direction: column;
+  gap: 15px;
+  .password {
+    background-color: #fff;
+    display: flex;
+    align-items: center;
+    border: 2px solid #000;
+    border-radius: 7px;
+
+    input {
+      border: none;
+    }
+
+    div {
+      padding: 5px;
+      height: 20px;
+      margin-top: -7px;
+    }
+  }
+  input {
+    padding: 15px;
+    width: 300px;
+    font-size: 16px;
+    outline: none;
+    border-radius: 7px;
+    width: 100%;
+  }
+}
